--- api/src/routes/manager.ts
@@ -1,10 +1,0 @@
-import express from "express"
-import { deleteTrainer, getTrainerDetails } from "src/controllers/manager";
-
-
-const router = express.Router()
-
-router.get("/", getTrainerDetails);
-router.delete("/:id", deleteTrainer); 
-
-export default router
