+++ api/src/server.ts
@@ -2,7 +2,7 @@ import express from 'express';
 import bodyParser from 'body-parser';
 import homeRoutes from './routes/home';
 import membersRoutes from "./routes/members";
-import managerRoutes from "./routes/manager"
+import trainerRoutes from "./routes/trainer";
 const cors = require("cors"); 
 
 const app = express();
@@ -16,7 +16,7 @@ app.use(bodyParser.urlencoded({extended:true}));
 
 app.use('/', homeRoutes);
 app.use("/users", membersRoutes);
-app.use('/staff?type=trainer',managerRoutes)
+app.use('/trainer/',trainerRoutes)
 
 app.listen(8081, () =>{
     console.log('server is running on port 8081.');
